Fix camera distortion issue on Galaxy S6
Calling startCaptureSession() twice fixes a vertically-stretched
preview issue from occurring. This does not occur on any Nexus
phones and I suspect it occurs on other Samsung phones as well.

I don't know why or how...and one day I hope to figure it all out.
But today is not that day.